created base_admin for all admin
