+++ core/exam/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.utils.translation import gettext_lazy as _
-from import_export.admin import ImportExportModelAdmin
+from utils.base_admin import BaseModelAdmin
 from import_export import resources
 from exam.models import *
 
@@ -11,7 +11,7 @@ class ExamTitleResource(resources.ModelResource):
         model = ExamTitle
 
 
-class ExamTitleModelAdmin(ImportExportModelAdmin):
+class ExamTitleModelAdmin(BaseModelAdmin):
     # form = TopicAdminForm
     list_display = ['name', 'expansion_of_name', 'is_published']
     sortable_by = ('name', 'expansion_of_name')
@@ -34,7 +34,6 @@ class ExamTitleModelAdmin(ImportExportModelAdmin):
     # paginator = Paginator
     # raw_id_fields = ('subject', 'chapter')
     # autocomplete_fields = ()
-    actions = ('set_item_to_published', 'set_item_to_unpublished', 'set_item_to_active', 'set_item_to_unactive')
     preserve_filters = True
     search_fields = ('name', 'expansion_of_name')
     readonly_fields = ('created_at', 'created_by', 'updated_at', 'updated_by')
@@ -53,12 +52,6 @@ class ExamTitleModelAdmin(ImportExportModelAdmin):
     class Meta:
         model = ExamTitle
 
-    def get_queryset(self, request):
-        qs = super().get_queryset(request)
-        if request.user.is_superuser:
-            return qs
-        return qs.filter(is_active=True)
-
     # def formfield_for_foreignkey(self, db_field, request, **kwargs):
     #     print(db_field.name, kwargs)
     #     if db_field.name == "subject":
@@ -66,28 +59,11 @@ class ExamTitleModelAdmin(ImportExportModelAdmin):
     #         print(db_field, request)
     #     return super().formfield_for_foreignkey(db_field, request, **kwargs)
 
-    def has_delete_permission(self, request, obj=None):
-        if request.user.is_superuser:
-            return True
-        return False
-
-    def get_list_filter(self, request):
-        filters = ['is_published', 'is_active', 'updated_at']
-        if not request.user.is_superuser:
-            filters.remove('is_active')
-        return filters
-
     # def get_exclude(self, request):
     #     excludes = super().get_exclude(request)
     #     if not request.user.is_superuser:
     #         excludes.remove('is_active')
     #     return excludes
-
-    # def get_actions(self, request):
-    #     actions = super().get_actions(request)
-    #     # if request.user.has_perm('exam.delete_exam_title'):
-    #     #     del actions['set_item_to_active']
-    #     return actions
 
     # def get_list_display(self, request):
     #     actions = super().get_list_display(request)
@@ -95,34 +71,6 @@ class ExamTitleModelAdmin(ImportExportModelAdmin):
     #         del actions['published_questions']
     #         del actions['unpublished_questions']
     #     return actions
-
-    def save_model(self, request, obj, form, change):
-        if not change:
-            obj.created_by = request.user
-        else:
-            obj.updated_by = request.user
-        obj.save()
-
-    def set_item_to_published(self, request, queryset):
-        count = queryset.update(is_published=True)
-        self.message_user(request, '{} items have been published successfully.'.format(count))
-
-    def set_item_to_unpublished(self, request, queryset):
-        count = queryset.update(is_published=True)
-        self.message_user(request, '{} items have been published successfully.'.format(count))
-
-    def set_item_to_active(self, request, queryset):
-        count = queryset.update(is_active=True)
-        self.message_user(request, '{} items have been published successfully.'.format(count))
-
-    def set_item_to_unactive(self, request, queryset):
-        count = queryset.update(is_active=True)
-        self.message_user(request, '{} items have been published successfully.'.format(count))
-
-    set_item_to_published.short_description = 'Publishing selected items'
-    set_item_to_unpublished.short_description = 'Unpublishing selected items'
-    set_item_to_active.short_description = 'Activating selected items'
-    set_item_to_unactive.short_description = 'Inactive/Delete selected items'
 
 
 admin.site.register(ExamTitle, ExamTitleModelAdmin)
